Adopt Sam's workflow for settings filters.
